modifying the structure of the project (it's hard to make 2 jobs working if they stand in separated folders); processor or writer or tasklet code of the 2 jobs are all now at the same level (the start of the application, because the order of the beans scanning is hard to manage)
