+++ src/main/resources/admin_suppressionmasse.txt
@@ -1,3 +1,3 @@
-230250
-230251
-100117
+213668
+158144
+213645
