+++ src/main/resources/sh/admin_suppressiondoublonnntportail.sh
@@ -1,5 +1,4 @@
 
-
 cd /home/batch/thesesBatchDelete/current
 
 /usr/java/jdk11/bin/java -Djava.security.egd=file:///dev/urandom -jar thesesBatchDelete.jar duplicatesNntPortailToDeleteJobBean >> ../logs/doublonnntportail_single.log &
